# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/assessment-periods": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the distinct assessment periods that have submitted health check data, for populating dropdowns.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health-checks"
                ],
                "summary": "List distinct assessment periods",
                "responses": {
                    "200": {
                        "description": "periods",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch assessment periods",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/teams/{teamId}/submission-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns how many team members have submitted a health check for the given assessment period, and whether a post-workshop survey exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health-checks"
                ],
                "summary": "Get a team's submission status for a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Assessment period (e.g. '2024 - 1st Half')",
                        "name": "assessmentPeriod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamSubmissionStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Missing teamId or assessmentPeriod",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get submission status",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamSubmissionStatusResponse": {
            "type": "object",
            "properties": {
                "allSubmitted": {
                    "type": "boolean"
                },
                "assessmentPeriod": {
                    "type": "string"
                },
                "postWorkshopExists": {
                    "type": "boolean"
                },
                "submittedMembers": {
                    "type": "integer"
                },
                "teamId": {
                    "type": "string"
                },
                "totalMembers": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}